db-model: adjust model
